src/DataManagementSystem/requirements.txt
argon2-cffi==21.1.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
Django==3.2.5
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
Pillow==8.4.0
protobuf==3.19.1
psycopg2-binary
pycparser==2.20
python-dateutil==2.8.2
pytz
six==1.16.0
sqlparse
text-unidecode==1.3
typing-extensions
wincertstore==0.2
python-dotenv
wget
pandas